+++ crates/pallets/starknet/src/simulations.rs
@@ -110,6 +110,39 @@ impl<T: Config> Pallet<T> {
         Ok(tx_execution_results)
     }
 
+    pub fn simulate_message(
+        message: HandleL1MessageTransaction,
+        simulation_flags: &SimulationFlags,
+    ) -> Result<Result<TransactionExecutionInfo, PlaceHolderErrorTypeForFailedStarknetExecution>, DispatchError> {
+        storage::transactional::with_transaction(|| {
+            storage::TransactionOutcome::Rollback(Result::<_, DispatchError>::Ok(Self::simulate_message_inner(
+                message,
+                simulation_flags,
+            )))
+        })
+        .map_err(|_| Error::<T>::FailedToCreateATransactionalStorageExecution)?
+    }
+
+    fn simulate_message_inner(
+        message: HandleL1MessageTransaction,
+        simulation_flags: &SimulationFlags,
+    ) -> Result<Result<TransactionExecutionInfo, PlaceHolderErrorTypeForFailedStarknetExecution>, DispatchError> {
+        let chain_id = Self::chain_id();
+        let block_context = Self::get_block_context();
+        let mut execution_config =
+            RuntimeExecutionConfigBuilder::new::<T>().with_simulation_mode(simulation_flags).build();
+
+        // Follow `offset` from Pallet Starknet where it is set to false
+        execution_config.set_offset_version(false);
+        let tx_execution_result =
+            Self::execute_message(message, chain_id, &block_context, &execution_config).map_err(|e| {
+                log::error!("Transaction execution failed during simulation: {e}");
+                PlaceHolderErrorTypeForFailedStarknetExecution
+            });
+
+        Ok(tx_execution_result)
+    }
+
     pub fn estimate_message_fee(message: HandleL1MessageTransaction) -> Result<(u128, u64, u64), DispatchError> {
         storage::transactional::with_transaction(|| {
             storage::TransactionOutcome::Rollback(Result::<_, DispatchError>::Ok(Self::estimate_message_fee_inner(
@@ -122,7 +155,8 @@ impl<T: Config> Pallet<T> {
     fn estimate_message_fee_inner(message: HandleL1MessageTransaction) -> Result<(u128, u64, u64), DispatchError> {
         let chain_id = Self::chain_id();
 
-        let tx_execution_infos = match message.into_executable::<T::SystemHash>(chain_id, Fee(u128::MAX), true).execute(
+        // Follow `offset` from Pallet Starknet where it is set to false
+        let tx_execution_infos = match message.into_executable::<T::SystemHash>(chain_id, Fee(u128::MAX), false).execute(
             &mut BlockifierStateAdapter::<T>::default(),
             &Self::get_block_context(),
             &RuntimeExecutionConfigBuilder::new::<T>().with_query_mode().with_disable_nonce_validation().build(),
@@ -239,5 +273,16 @@ impl<T: Config> Pallet<T> {
                 executable.execute(&mut BlockifierStateAdapter::<T>::default(), block_context, execution_config)
             }
         }
+    }
+
+    fn execute_message(
+        message: HandleL1MessageTransaction,
+        chain_id: Felt252Wrapper,
+        block_context: &BlockContext,
+        execution_config: &ExecutionConfig,
+    ) -> Result<TransactionExecutionInfo, TransactionExecutionError> {
+        // Follow `offset` from Pallet Starknet where it is set to false
+        let executable = message.into_executable::<T::SystemHash>(chain_id, Fee::default(), false);
+        executable.execute(&mut BlockifierStateAdapter::<T>::default(), block_context, execution_config)
     }
 }